@@ -1,211 +1,291 @@
 {
-  "name": "@material-ui/workspace",
+  "name": "@mui/monorepo",
+  "version": "6.0.0-rc.0",
   "private": true,
-  "author": "Material-UI Team",
-  "version": "3.9.0",
-  "description": "Material-UI's workspace package",
-  "keywords": [
-    "react",
-    "react-component",
-    "material design",
-    "material-ui"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/mui-org/material-ui.git"
-  },
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/mui-org/material-ui/issues"
-  },
-  "homepage": "https://material-ui.com/",
   "scripts": {
-    "argos": "argos upload test/regressions/screenshots/chrome --token $ARGOS_TOKEN",
-    "docs:api": "yarn docs:api:core && yarn docs:api:lab",
-    "docs:api:core": "rimraf ./pages/api && cross-env BABEL_ENV=docs-development babel-node ./docs/scripts/buildApi.js ./packages/material-ui/src ./pages/api",
-    "docs:api:lab": "rimraf ./pages/lab/api && cross-env BABEL_ENV=docs-development babel-node ./docs/scripts/buildApi.js ./packages/material-ui-lab/src ./pages/lab/api",
-    "docs:build": "rimraf .next && cross-env NODE_ENV=production BABEL_ENV=docs-production next build",
-    "docs:build-sw": "babel-node ./docs/scripts/buildServiceWorker.js",
-    "docs:deploy": "git push material-ui-docs master:latest",
-    "docs:dev": "rimraf node_modules/.cache/babel-loader && cross-env BABEL_ENV=docs-development next dev",
-    "docs:export": "rimraf docs/export && next export -o docs/export && yarn docs:build-sw && cp -r docs/static/. docs/export",
-    "docs:icons": "rimraf static/icons/* && babel-node ./docs/scripts/buildIcons.js",
-    "docs:size-why": "DOCS_STATS_ENABLED=true yarn docs:build",
-    "docs:start": "next start",
-    "jsonlint": "yarn --silent jsonlint:files | xargs -n1 jsonlint -q -c && echo \"jsonlint: no lint errors\"",
-    "jsonlint:files": "find . -name \"*.json\" | grep -v -f .eslintignore",
-    "lint": "eslint . --cache --report-unused-disable-directives && echo \"eslint: no lint errors\"",
-    "lint:fix": "eslint . --cache --fix && echo \"eslint: no lint errors\"",
-    "prettier": "yarn babel-node ./scripts/prettier.js",
-    "prettier:all": "yarn babel-node ./scripts/prettier.js write",
-    "size": "size-limit",
-    "size:overhead:why": "size-limit --why ./test/size/overhead.js",
-    "size:why": "size-limit --why packages/material-ui/build/index.js",
-    "start": "yarn docs:dev",
-    "test": "yarn lint && yarn typescript && yarn test:coverage",
-    "test:coverage": "cross-env NODE_ENV=test BABEL_ENV=coverage nyc mocha 'packages/material-ui/**/*.test.js' 'packages/material-ui-utils/**/*.test.js' 'packages/material-ui-styles/**/*.test.js' --exclude '**/node_modules/**' && nyc report -r lcovonly",
-    "test:coverage:html": "cross-env NODE_ENV=test BABEL_ENV=coverage nyc mocha 'packages/material-ui/**/*.test.js' 'packages/material-ui-utils/**/*.test.js' 'packages/material-ui-styles/**/*.test.js' --exclude '**/node_modules/**' && nyc report --reporter=html",
-    "test:karma": "cross-env NODE_ENV=test karma start test/karma.conf.js --single-run",
-    "test:regressions": "webpack --config test/regressions/webpack.config.js && rimraf test/regressions/screenshots/chrome/* && vrtest run --config test/vrtest.config.js --record",
-    "test:umd": "yarn babel-node packages/material-ui/test/umd/run.js",
-    "test:unit": "cross-env NODE_ENV=test mocha 'packages/**/*.test.js' 'docs/**/*.test.js' --exclude '**/node_modules/**'",
-    "test:watch": "yarn test:unit --watch",
-    "typescript": "lerna run typescript"
+    "preinstall": "npx only-allow pnpm",
+    "proptypes": "tsx ./scripts/generateProptypes.ts",
+    "deduplicate": "pnpm dedupe",
+    "benchmark:browser": "pnpm --filter benchmark browser",
+    "build": "lerna run build --ignore docs",
+    "build:ci": "lerna run build --ignore docs --concurrency 8 --skip-nx-cache",
+    "build:public": "lerna run --no-private build",
+    "build:public:ci": "lerna run --no-private build --concurrency 8 --skip-nx-cache",
+    "build:codesandbox": "NODE_OPTIONS=\"--max_old_space_size=4096\" lerna run --concurrency 8 --scope \"@mui/*\" --scope \"@mui-internal/*\" --no-private build",
+    "release:version": "lerna version --no-changelog --no-push --no-git-tag-version --no-private --force-publish=@mui/core-downloads-tracker",
+    "release:build": "lerna run --concurrency 8 --no-private build --skip-nx-cache",
+    "release:changelog": "node scripts/releaseChangelog.mjs",
+    "release:publish": "pnpm publish --recursive --tag next",
+    "release:publish:dry-run": "pnpm publish --recursive --tag next --registry=\"http://localhost:4873/\"",
+    "release:tag": "node scripts/releaseTag.mjs",
+    "docs:api": "rimraf --glob ./docs/pages/**/api-docs ./docs/pages/**/api && pnpm docs:api:build",
+    "docs:api:build": "tsx ./scripts/buidApiDocs/index.ts",
+    "docs:build": "pnpm --filter docs build",
+    "docs:build-sw": "pnpm --filter docs build-sw",
+    "docs:build-color-preview": "babel-node scripts/buildColorTypes",
+    "docs:deploy": "pnpm --filter docs run deploy",
+    "docs:dev": "pnpm --filter docs dev",
+    "docs:icons": "pnpm --filter docs icons",
+    "docs:size-why": "cross-env DOCS_STATS_ENABLED=true pnpm docs:build",
+    "docs:start": "pnpm --filter docs start",
+    "docs:create-playground": "pnpm --filter docs create-playground",
+    "docs:i18n": "cross-env BABEL_ENV=development babel-node --extensions \".tsx,.ts,.js\" ./docs/scripts/i18n.ts",
+    "docs:link-check": "pnpm --filter docs link-check",
+    "docs:typescript": "pnpm docs:typescript:formatted --watch",
+    "docs:typescript:check": "pnpm --filter docs typescript",
+    "docs:typescript:formatted": "tsx ./docs/scripts/formattedTSDemos",
+    "docs:mdicons:synonyms": "cross-env BABEL_ENV=development babel-node --extensions \".tsx,.ts,.js,.mjs\" ./docs/scripts/updateIconSynonyms && pnpm prettier",
+    "docs:zipRules": "cd docs && rm mui-vale.zip && zip -r mui-vale.zip mui-vale && cd ../ && vale sync",
+    "extract-error-codes": "cross-env MUI_EXTRACT_ERROR_CODES=true lerna run --concurrency 8 build:modern",
+    "rsc:build": "tsx ./packages/rsc-builder/buildRsc.ts",
+    "template:screenshot": "cross-env BABEL_ENV=development babel-node --extensions \".tsx,.ts,.js\" ./docs/scripts/generateTemplateScreenshots",
+    "template:update-theme": "cross-env BABEL_ENV=development babel-node --extensions \".tsx,.ts,.js\" ./docs/scripts/updateTemplatesTheme",
+    "install:codesandbox": "pnpm install --no-frozen-lockfile",
+    "jsonlint": "node ./scripts/jsonlint.mjs",
+    "eslint": "eslint . --cache --report-unused-disable-directives --ext .js,.ts,.tsx --max-warnings 0",
+    "eslint:ci": "eslint . --report-unused-disable-directives --ext .js,.ts,.tsx --max-warnings 0",
+    "stylelint": "stylelint --reportInvalidScopeDisables --reportNeedlessDisables \"docs/**/*.{js,ts,tsx}\" --ignore-path .gitignore",
+    "markdownlint": "markdownlint-cli2 \"**/*.md\"",
+    "valelint": "git ls-files | grep -h \".md$\" | xargs vale --filter='.Level==\"error\"'",
+    "prettier": "pretty-quick --ignore-path .eslintignore --branch next",
+    "prettier:all": "prettier --write . --ignore-path .eslintignore",
+    "size:snapshot": "node --max-old-space-size=4096 ./scripts/sizeSnapshot/create",
+    "size:why": "pnpm size:snapshot --analyze",
+    "start": "pnpm install && pnpm docs:dev",
+    "test": "node scripts/test.mjs",
+    "tc": "nx run nx_test_tc",
+    "test:extended": "pnpm eslint && pnpm typescript && pnpm test:coverage",
+    "test:coverage": "nx run nx_test_coverage",
+    "test:coverage:ci": "nx run nx_test_coverage_ci",
+    "test:coverage:html": "nx run nx_test_coverage_html",
+    "test:e2e": "cross-env NODE_ENV=production pnpm test:e2e:build && concurrently --success first --kill-others \"pnpm test:e2e:run\" \"pnpm test:e2e:server\"",
+    "test:e2e:build": "nx run nx_test_e2e_build",
+    "test:e2e:dev": "concurrently \"pnpm test:e2e:build --watch\" \"pnpm test:e2e:server\"",
+    "test:e2e:run": "nx run nx_test_e2e_run",
+    "test:e2e:server": "serve test/e2e -p 5001",
+    "test:e2e-website": "playwright test test/e2e-website --config test/e2e-website/playwright.config.ts",
+    "test:e2e-website:dev": "cross-env PLAYWRIGHT_TEST_BASE_URL=http://localhost:3000 playwright test test/e2e-website --config test/e2e-website/playwright.config.ts",
+    "test:karma": "nx run nx_test_karma",
+    "test:karma:profile": "nx run nx_test_karma_profile",
+    "test:regressions": "cross-env NODE_ENV=production pnpm test:regressions:build && concurrently --success first --kill-others \"pnpm test:regressions:run\" \"pnpm test:regressions:server\"",
+    "test:regressions:build": "webpack --config test/regressions/webpack.config.js",
+    "test:regressions:dev": "concurrently \"pnpm test:regressions:build --watch\" \"pnpm test:regressions:server\"",
+    "test:regressions:run": "nx run nx_test_regressions_run",
+    "test:regressions:server": "serve test/regressions -p 5001",
+    "test:unit": "nx run nx_test_unit",
+    "test:argos": "node ./scripts/pushArgos.mjs",
+    "typescript": "lerna run --no-bail --parallel typescript",
+    "typescript:ci": "lerna run --concurrency 3 --no-bail --no-sort typescript",
+    "use-react-version": "node ./scripts/useReactVersion.mjs",
+    "validate-declarations": "tsx scripts/validateTypescriptDeclarations.mts",
+    "generate-codeowners": "node scripts/generateCodeowners.mjs",
+    "canary:release": "tsx ./scripts/canaryRelease.mts",
+    "nx_test_tc": "node test/cli.js",
+    "nx_test_coverage_ci": "cross-env NODE_ENV=test BABEL_ENV=coverage nyc --reporter=lcov mocha 'packages/**/*.test.{js,ts,tsx}' 'packages-internal/**/*.test.{js,ts,tsx}' 'docs/**/*.test.{js,ts,tsx}'",
+    "nx_test_coverage_html": "cross-env NODE_ENV=test BABEL_ENV=coverage nyc --reporter=html mocha 'packages/**/*.test.{js,ts,tsx}' 'packages-internal/**/*.test.{js,ts,tsx}' 'docs/**/*.test.{js,ts,tsx}'",
+    "nx_test_coverage": "cross-env NODE_ENV=test BABEL_ENV=coverage nyc --reporter=text mocha 'packages/**/*.test.{js,ts,tsx}' 'packages-internal/**/*.test.{js,ts,tsx}' 'docs/**/*.test.{js,ts,tsx}'",
+    "nx_test_e2e_build": "webpack --config test/e2e/webpack.config.js",
+    "nx_test_e2e_run": "mocha --config test/e2e/.mocharc.js 'test/e2e/**/*.test.{js,ts,tsx}'",
+    "nx_test_karma_profile": "cross-env NODE_ENV=test karma start test/karma.conf.profile.js",
+    "nx_test_karma": "cross-env NODE_ENV=test karma start test/karma.conf.js",
+    "nx_test_regressions_run": "mocha --config test/regressions/.mocharc.js --delay 'test/regressions/**/*.test.js'",
+    "nx_test_unit": "cross-env NODE_ENV=test mocha 'packages/**/*.test.{js,ts,tsx}' 'packages-internal/**/*.test.{js,ts,tsx}' 'docs/**/*.test.{js,ts,tsx}'"
   },
-  "peerDependencies": {
-    "react": "*",
-    "react-dom": "*"
+  "dependencies": {
+    "@googleapis/sheets": "^9.1.0",
+    "@netlify/functions": "^2.8.1",
+    "@slack/bolt": "^3.21.1",
+    "execa": "^9.3.1",
+    "google-auth-library": "^9.13.0"
   },
   "devDependencies": {
-    "@babel/cli": "^7.2.0",
-    "@babel/core": "^7.2.2",
-    "@babel/node": "^7.2.2",
-    "@babel/plugin-proposal-class-properties": "^7.2.1",
-    "@babel/plugin-proposal-object-rest-spread": "^7.2.0",
-    "@babel/plugin-transform-object-assign": "^7.2.0",
-    "@babel/plugin-transform-runtime": "^7.2.0",
-    "@babel/preset-env": "^7.2.0",
-    "@babel/preset-react": "^7.0.0",
-    "@babel/register": "^7.0.0",
-    "@date-io/date-fns": "^1.0.0",
-    "@emotion/core": "^10.0.0",
-    "@emotion/styled": "^10.0.0",
-    "@types/enzyme": "^3.1.4",
-    "@types/react": "^16.7.10",
-    "@weco/next-plugin-transpile-modules": "0.0.2",
-    "accept-language": "^3.0.18",
-    "argos-cli": "^0.0.9",
-    "autoprefixer": "^9.0.0",
-    "autosuggest-highlight": "^3.1.1",
-    "babel-eslint": "^10.0.0",
-    "babel-loader": "^8.0.0",
-    "babel-plugin-istanbul": "^5.0.0",
-    "babel-plugin-module-resolver": "^3.0.0",
-    "babel-plugin-preval": "^2.0.0",
-    "babel-plugin-react-remove-properties": "^0.2.5",
-    "babel-plugin-transform-dev-warning": "^0.1.0",
-    "babel-plugin-transform-react-constant-elements": "^6.23.0",
-    "babel-plugin-transform-react-remove-prop-types": "^0.4.21",
-    "chai": "^4.1.2",
-    "clean-css": "^4.1.11",
-    "clipboard-copy": "^2.0.0",
-    "cross-env": "^5.1.1",
-    "css-loader": "^2.0.0",
-    "css-mediaquery": "^0.1.2",
-    "date-fns": "2.0.0-alpha.21",
-    "doctrine": "^3.0.0",
-    "downshift": "^3.0.0",
-    "dtslint": "^0.3.0",
-    "emotion-theming": "^10.0.5",
-    "enzyme": "^3.5.0",
-    "enzyme-adapter-react-16": "^1.3.0",
-    "eslint": "^5.9.0",
-    "eslint-config-airbnb": "^17.0.0",
-    "eslint-config-prettier": "^3.3.0",
-    "eslint-import-resolver-webpack": "^0.10.0",
-    "eslint-plugin-babel": "^5.3.0",
-    "eslint-plugin-import": "^2.8.0",
-    "eslint-plugin-jsx-a11y": "^6.0.3",
-    "eslint-plugin-material-ui": "file:packages/eslint-plugin-material-ui",
-    "eslint-plugin-mocha": "^5.0.0",
-    "eslint-plugin-react": "^7.4.0",
-    "expect-puppeteer": "^3.5.1",
-    "fg-loadcss": "^2.0.1",
-    "file-loader": "^2.0.0",
-    "final-form": "^4.11.0",
-    "fs-extra": "^7.0.0",
-    "glob": "^7.1.2",
-    "glob-gitignore": "^1.0.11",
-    "gm": "^1.23.0",
-    "isomorphic-fetch": "^2.2.1",
-    "jsdom": "^13.0.0",
-    "jsonlint": "^1.6.3",
-    "jss-plugin-template": "^10.0.0-alpha.3",
-    "jss-rtl": "^0.2.1",
-    "karma": "3.0.0",
-    "karma-browserstack-launcher": "^1.3.0",
-    "karma-chrome-launcher": "^2.2.0",
-    "karma-mocha": "^1.3.0",
-    "karma-mocha-reporter": "^2.2.5",
-    "karma-sourcemap-loader": "^0.3.7",
-    "karma-webpack": "^3.0.0",
-    "lerna": "^3.4.3",
-    "lz-string": "^1.4.4",
-    "markdown-to-jsx": "^6.8.3",
-    "material-ui-pickers": "^2.0.2",
-    "material-ui-popup-state": "^1.0.1",
-    "mocha": "^5.0.0",
-    "next": "^7.0.0",
-    "notistack": "^0.4.0",
-    "nyc": "^13.0.0",
-    "postcss": "^7.0.0",
-    "prettier": "^1.8.2",
-    "puppeteer": "^1.5.0",
-    "raw-loader": "^0.5.1",
-    "react": "next",
-    "react-autosuggest": "^9.3.2",
-    "react-docgen": "^3.0.0-beta10",
-    "react-dom": "next",
-    "react-draggable": "^3.0.5",
-    "react-final-form": "^4.0.2",
-    "react-frame-component": "^4.0.2",
-    "react-inspector": "^2.2.2",
-    "react-jss": "^8.6.1",
-    "react-number-format": "^4.0.0",
-    "react-redux": "^6.0.0",
-    "react-router": "^4.2.0",
-    "react-router-dom": "^4.2.2",
-    "react-select": "^2.0.0",
-    "react-swipeable-views": "^0.13.0",
-    "react-test-renderer": "^16.1.1",
-    "react-text-mask": "^5.0.2",
-    "react-virtualized": "^9.21.0",
-    "recast": "^0.16.0",
-    "recharts": "^1.1.0",
-    "redux": "^4.0.0",
-    "redux-logger": "^3.0.6",
-    "rimraf": "^2.6.2",
-    "rollup": "^0.67.0",
-    "rollup-plugin-babel": "^4.0.0-beta.8",
-    "rollup-plugin-commonjs": "^9.1.3",
-    "rollup-plugin-node-globals": "^1.4.0",
-    "rollup-plugin-node-resolve": "^3.3.0",
-    "rollup-plugin-size-snapshot": "^0.7.0",
-    "rollup-plugin-uglify": "^6.0.0",
-    "sinon": "^7.0.0",
-    "size-limit": "^0.21.0",
-    "styled-components": "^4.1.1",
-    "tslint": "^5.11.0",
-    "typescript": "3.1.6",
-    "url-loader": "^1.0.1",
-    "vrtest": "^0.2.0",
-    "webfontloader": "^1.6.28",
-    "webpack": "4.20.2",
-    "webpack-bundle-analyzer": "^3.0.0",
-    "webpack-cli": "^3.1.0"
+    "@argos-ci/core": "^2.4.1",
+    "@babel/cli": "^7.24.8",
+    "@babel/core": "^7.25.2",
+    "@babel/node": "^7.25.0",
+    "@babel/plugin-transform-react-constant-elements": "^7.25.1",
+    "@babel/plugin-transform-runtime": "^7.24.7",
+    "@babel/preset-env": "^7.25.3",
+    "@babel/preset-react": "^7.24.7",
+    "@babel/preset-typescript": "^7.24.7",
+    "@babel/register": "^7.24.6",
+    "@mui-internal/api-docs-builder": "workspace:^",
+    "@mui-internal/api-docs-builder-core": "workspace:^",
+    "@mui/internal-docs-utils": "workspace:^",
+    "@mui/internal-scripts": "workspace:^",
+    "@mui/internal-test-utils": "workspace:^",
+    "@mui/joy": "workspace:*",
+    "@mui/material": "workspace:^",
+    "@mui/utils": "workspace:^",
+    "@next/eslint-plugin-next": "^14.2.5",
+    "@octokit/rest": "^21.0.2",
+    "@pigment-css/react": "0.0.20",
+    "@playwright/test": "1.46.1",
+    "@types/babel__core": "^7.20.5",
+    "@types/fs-extra": "^11.0.4",
+    "@types/lodash": "^4.17.7",
+    "@types/mocha": "^10.0.7",
+    "@types/node": "^18.19.44",
+    "@types/react": "^18.3.3",
+    "@types/yargs": "^17.0.33",
+    "@typescript-eslint/eslint-plugin": "^7.18.0",
+    "@typescript-eslint/parser": "^7.18.0",
+    "babel-loader": "^9.1.3",
+    "babel-plugin-istanbul": "^7.0.0",
+    "babel-plugin-macros": "^3.1.0",
+    "babel-plugin-module-resolver": "^5.0.2",
+    "babel-plugin-optimize-clsx": "^2.6.2",
+    "babel-plugin-react-remove-properties": "^0.3.0",
+    "babel-plugin-transform-inline-environment-variables": "^0.4.4",
+    "babel-plugin-transform-react-remove-prop-types": "^0.4.24",
+    "chalk": "^5.3.0",
+    "compression-webpack-plugin": "^11.1.0",
+    "concurrently": "^8.2.2",
+    "cpy-cli": "^5.0.0",
+    "cross-env": "^7.0.3",
+    "danger": "^12.3.3",
+    "eslint": "^8.57.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-config-airbnb-typescript": "^18.0.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-import-resolver-webpack": "^0.13.8",
+    "eslint-plugin-babel": "^5.3.1",
+    "eslint-plugin-filenames": "^1.3.2",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-jsx-a11y": "6.7.1",
+    "eslint-plugin-material-ui": "workspace:^",
+    "eslint-plugin-mocha": "^10.5.0",
+    "eslint-plugin-react": "^7.35.0",
+    "eslint-plugin-react-compiler": "0.0.0-experimental-9aef357-20240815",
+    "eslint-plugin-react-hooks": "^4.6.2",
+    "fast-glob": "^3.3.2",
+    "fs-extra": "^11.2.0",
+    "globby": "^14.0.2",
+    "karma": "^6.4.4",
+    "karma-browserstack-launcher": "~1.6.0",
+    "karma-chrome-launcher": "^3.2.0",
+    "karma-coverage-istanbul-reporter": "^3.0.3",
+    "karma-firefox-launcher": "^2.1.3",
+    "karma-mocha": "^2.0.1",
+    "karma-sourcemap-loader": "^0.4.0",
+    "karma-webpack": "^5.0.0",
+    "lerna": "^8.1.8",
+    "lodash": "^4.17.21",
+    "markdownlint-cli2": "^0.13.0",
+    "mocha": "^10.7.3",
+    "nx": "^19.6.0",
+    "nyc": "^17.0.0",
+    "piscina": "^4.6.1",
+    "postcss-styled-syntax": "^0.6.4",
+    "prettier": "^3.3.3",
+    "pretty-quick": "^4.0.0",
+    "process": "^0.11.10",
+    "rimraf": "^5.0.10",
+    "serve": "^14.2.3",
+    "stylelint": "^15.11.0",
+    "stylelint-config-standard": "^34.0.0",
+    "stylelint-processor-styled-components": "^1.10.0",
+    "terser-webpack-plugin": "^5.3.10",
+    "tsx": "^4.17.0",
+    "typescript": "^5.5.4",
+    "webpack": "^5.93.0",
+    "webpack-bundle-analyzer": "^4.10.2",
+    "webpack-cli": "^5.1.4",
+    "yargs": "^17.7.2"
+  },
+  "packageManager": "pnpm@9.7.1",
+  "engines": {
+    "pnpm": "9.7.1"
   },
   "resolutions": {
-    "**/hoist-non-react-statics": "^3.2.1",
-    "react-jss/jss": "^10.0.0-alpha.3"
+    "@babel/core": "^7.25.2",
+    "@babel/code-frame": "^7.24.7",
+    "@babel/plugin-transform-runtime": "^7.24.7",
+    "@babel/preset-env": "^7.25.3",
+    "@babel/preset-react": "^7.24.7",
+    "@babel/preset-typescript": "^7.24.7",
+    "@babel/runtime": "^7.25.0",
+    "@babel/types": "^7.25.2",
+    "@definitelytyped/header-parser": "^0.2.12",
+    "@definitelytyped/typescript-versions": "^0.1.4",
+    "@definitelytyped/utils": "^0.1.7",
+    "@types/node": "^18.19.44",
+    "@types/react": "18.3.3",
+    "@types/react-dom": "18.3.0",
+    "cross-fetch": "^4.0.0",
+    "@pigment-css/react": "0.0.20",
+    "@pigment-css/unplugin": "0.0.20",
+    "@pigment-css/nextjs-plugin": "0.0.20",
+    "@pigment-css/vite-plugin": "0.0.20"
   },
-  "sideEffects": false,
   "nyc": {
     "include": [
-      "packages/material-ui/src/**/*.js",
-      "packages/material-ui/utils/**/*.js",
-      "packages/material-ui/styles/**/*.js"
+      "packages/mui*/src/**/*.{js,ts,tsx}"
     ],
     "exclude": [
-      "**/*.test.js"
+      "**/*.test.{js,ts,tsx}",
+      "**/*.test/*"
     ],
     "sourceMap": false,
     "instrument": false
   },
-  "engines": {
-    "node": ">=8.0.0"
+  "nx": {
+    "targets": {
+      "nx_test_tc": {
+        "dependsOn": [
+          "@mui/internal-test-utils:build"
+        ]
+      },
+      "nx_test_unit": {
+        "dependsOn": [
+          "@mui/internal-test-utils:build"
+        ]
+      },
+      "nx_test_karma": {
+        "dependsOn": [
+          "@mui/internal-test-utils:build"
+        ]
+      },
+      "nx_test_karma_profile": {
+        "dependsOn": [
+          "@mui/internal-test-utils:build"
+        ]
+      },
+      "nx_test_coverage": {
+        "dependsOn": [
+          "@mui/internal-test-utils:build"
+        ]
+      },
+      "nx_test_coverage_ci": {
+        "dependsOn": [
+          "@mui/internal-test-utils:build"
+        ]
+      },
+      "nx_test_coverage_html": {
+        "dependsOn": [
+          "@mui/internal-test-utils:build"
+        ]
+      },
+      "nx_test_regressions_run": {
+        "dependsOn": [
+          "@mui/internal-test-utils:build"
+        ]
+      },
+      "nx_test_e2e_build": {
+        "dependsOn": [
+          "@mui/internal-test-utils:build"
+        ]
+      },
+      "nx_test_e2e_run": {
+        "dependsOn": [
+          "@mui/internal-test-utils:build"
+        ]
+      }
+    }
   },
-  "workspaces": [
-    "packages/*"
-  ],
-  "dependencies": {}
+  "pnpm": {
+    "patchedDependencies": {
+      "styled-components@6.1.12": "patches/styled-components@6.1.12.patch"
+    }
+  }
 }
