@@ -1,61 +1,144 @@
-import React from 'react';
+import * as React from 'react';
+import { useTheme } from '@mui/material/styles';
+import useMediaQuery from '@mui/material/useMediaQuery';
+import { useRouter } from 'next/router';
+import { useNoSsrCodeVariant } from 'docs/src/modules/utils/codeVariant';
+import { useNoSsrCodeStyling } from 'docs/src/modules/utils/codeStylingSolution';
+import { useUserLanguage } from '@mui/docs/i18n';
+import { pathnameToLanguage } from 'docs/src/modules/utils/helpers';
 
+// So we can write code like:
+//
+// <Button
+//   data-ga-event-category="demo"
+//   data-ga-event-action="expand"
+// >
+//   Foo
+// </Button>
 function handleClick(event) {
-  const rootNode = document;
   let element = event.target;
 
-  while (element && element !== rootNode) {
+  while (element && element !== document) {
     const category = element.getAttribute('data-ga-event-category');
 
     // We reach a tracking element, no need to look higher in the dom tree.
     if (category) {
-      window.ga('send', {
-        hitType: 'event',
-        eventCategory: category,
+      const split = parseFloat(element.getAttribute('data-ga-event-split'));
+
+      if (split && split < Math.random()) {
+        return;
+      }
+
+      window.gtag('event', category, {
         eventAction: element.getAttribute('data-ga-event-action'),
         eventLabel: element.getAttribute('data-ga-event-label'),
       });
       break;
     }
 
-    element = element.parentNode;
+    element = element.parentElement;
   }
 }
 
-let binded = false;
+let boundDataGaListener = false;
 
-// So we can write code like:
-//
-// <Button
-//   ga-event-category="demo"
-//   ga-event-action="expand"
-// >
-//   Foo
-// </Button>
-function bindEvents() {
-  if (binded) {
-    return;
-  }
+/**
+ * basically just a `useAnalytics` hook.
+ * However, it needs the redux store which is created
+ * in the same component this "hook" is used.
+ */
+function GoogleAnalytics() {
+  React.useEffect(() => {
+    if (!boundDataGaListener) {
+      boundDataGaListener = true;
+      document.addEventListener('click', handleClick);
+    }
+  }, []);
+  const router = useRouter();
+  const timeout = React.useRef();
 
-  binded = true;
-  document.addEventListener('click', handleClick);
-}
-
-class GoogleAnalytics extends React.Component {
-  googleTimer = null;
-
-  componentDidMount() {
-    bindEvents();
+  React.useEffect(() => {
     // Wait for the title to be updated.
-    setTimeout(() => {
-      window.ga('set', { page: window.location.pathname });
-      window.ga('send', { hitType: 'pageview' });
-    });
-  }
+    // React fires useEffect twice in dev mode
+    clearTimeout(timeout.current);
+    timeout.current = setTimeout(() => {
+      const { canonicalAsServer } = pathnameToLanguage(window.location.pathname);
 
-  render() {
-    return null;
-  }
+      // https://developers.google.com/analytics/devguides/collection/ga4/views?client_type=gtag
+      window.gtag('event', 'page_view', {
+        page_title: document.title,
+        page_location: canonicalAsServer,
+        productId: document.querySelector('meta[name="mui:productId"]').content,
+        productCategoryId: document.querySelector('meta[name="mui:productCategoryId"]').content,
+      });
+    });
+  }, [router.route]);
+
+  const codeVariant = useNoSsrCodeVariant();
+  React.useEffect(() => {
+    window.gtag('set', 'user_properties', {
+      codeVariant,
+    });
+  }, [codeVariant]);
+
+  const userLanguage = useUserLanguage();
+  React.useEffect(() => {
+    window.gtag('set', 'user_properties', {
+      userLanguage,
+    });
+  }, [userLanguage]);
+
+  React.useEffect(() => {
+    /**
+     * Based on https://developer.mozilla.org/en-US/docs/Web/API/Window/devicePixelRatio#Monitoring_screen_resolution_or_zoom_level_changes
+     * Adjusted to track 3 or more different ratios
+     */
+    function trackDevicePixelRation() {
+      const devicePixelRatio = Math.round(window.devicePixelRatio * 10) / 10;
+      window.gtag('set', 'user_properties', {
+        devicePixelRatio,
+      });
+    }
+
+    trackDevicePixelRation();
+
+    /**
+     * @type {MediaQueryList}
+     */
+    const matchMedia = window.matchMedia(`(resolution: ${window.devicePixelRatio}dppx)`);
+    // Intentionally use deprecated listener methods to support iOS & old browsers
+    matchMedia.addListener(trackDevicePixelRation);
+    return () => {
+      matchMedia.removeListener(trackDevicePixelRation);
+    };
+  }, []);
+
+  const prefersDarkMode = useMediaQuery('(prefers-color-scheme: dark)', { noSsr: true });
+  const colorSchemeOS = prefersDarkMode ? 'dark' : 'light';
+
+  const theme = useTheme();
+  const colorScheme = theme.palette.mode;
+
+  React.useEffect(() => {
+    window.gtag('set', 'user_properties', {
+      colorSchemeOS,
+    });
+  }, [colorSchemeOS]);
+
+  React.useEffect(() => {
+    window.gtag('set', 'user_properties', {
+      colorScheme,
+    });
+  }, [colorScheme]);
+
+  const codeStylingVariant = useNoSsrCodeStyling();
+  React.useEffect(() => {
+    window.gtag('set', 'user_properties', {
+      codeStylingVariant,
+    });
+  }, [codeStylingVariant]);
+
+  return null;
 }
 
-export default GoogleAnalytics;
+export default React.memo(GoogleAnalytics);
