@@ -1,20 +1,8 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import { withStyles } from '@material-ui/core/styles';
-import MuiAppBar from '@material-ui/core/AppBar';
-
-const styles = theme => ({
-  root: {
-    color: theme.palette.common.white,
-  },
-});
+import * as React from 'react';
+import MuiAppBar from '@mui/material/AppBar';
 
 function AppBar(props) {
-  return <MuiAppBar elevation={0} position="static" {...props} />;
+  return <MuiAppBar elevation={0} position="fixed" {...props} />;
 }
 
-AppBar.propTypes = {
-  classes: PropTypes.object.isRequired,
-};
-
-export default withStyles(styles)(AppBar);
+export default AppBar;
