@@ -1,31 +1,42 @@
-import React from 'react';
-import { withStyles } from '@material-ui/core/styles';
-import MuiButton from '@material-ui/core/Button';
+import * as React from 'react';
+import { experimentalStyled as styled } from '@mui/material/styles';
+import MuiButton from '@mui/material/Button';
 
-const styles = theme => ({
-  root: {
-    borderRadius: 0,
-    fontWeight: theme.typography.fontWeightMedium,
-    fontFamily: theme.typography.fontFamilySecondary,
-    padding: `${theme.spacing.unit * 2 - 1}px ${theme.spacing.unit * 4}px`,
-    fontSize: theme.typography.pxToRem(14),
+const ButtonRoot = styled(MuiButton)(({ theme }) => ({
+  borderRadius: 0,
+  fontWeight: theme.typography.fontWeightMedium,
+  fontFamily: theme.typography.h1.fontFamily,
+  padding: theme.spacing(2, 4),
+  fontSize: theme.typography.pxToRem(14),
+  boxShadow: 'none',
+  '&:active, &:focus': {
     boxShadow: 'none',
-    '&:active, &:focus': {
-      boxShadow: 'none',
+  },
+  variants: [
+    {
+      props: {
+        size: 'small',
+      },
+      style: {
+        padding: theme.spacing(1, 3),
+        fontSize: theme.typography.pxToRem(13),
+      },
     },
-  },
-  sizeSmall: {
-    padding: `${theme.spacing.unit}px ${theme.spacing.unit * 3}px`,
-    fontSize: theme.typography.pxToRem(13),
-  },
-  sizeLarge: {
-    padding: `${theme.spacing.unit * 3 - 3}px ${theme.spacing.unit * 6}px`,
-    fontSize: theme.typography.pxToRem(16),
-  },
-});
+    {
+      props: {
+        size: 'large',
+      },
+      style: {
+        padding: theme.spacing(2, 5),
+        fontSize: theme.typography.pxToRem(16),
+      },
+    },
+  ],
+}));
 
+// See https://mui.com/guides/typescript/#usage-of-component-prop for why the types uses `C`.
 function Button(props) {
-  return <MuiButton {...props} />;
+  return <ButtonRoot {...props} />;
 }
 
-export default withStyles(styles)(Button);
+export default Button;
