@@ -1,33 +1,49 @@
-import React from 'react';
+import * as React from 'react';
 import ReactMarkdown from 'markdown-to-jsx';
-import { withStyles } from '@material-ui/core/styles';
-import Typography from '@material-ui/core/Typography';
-
-const styles = theme => ({
-  listItem: {
-    marginTop: theme.spacing.unit,
-  },
-});
+import Box from '@mui/material/Box';
+import Typography from '@mui/material/Typography';
+import Link from '@mui/material/Link';
 
 const options = {
   overrides: {
-    h1: { component: props => <Typography gutterBottom variant="h4" {...props} /> },
-    h2: { component: props => <Typography gutterBottom variant="h6" {...props} /> },
-    h3: { component: props => <Typography gutterBottom variant="subtitle1" {...props} /> },
-    h4: { component: props => <Typography gutterBottom variant="caption" paragraph {...props} /> },
-    p: { component: props => <Typography paragraph {...props} /> },
+    h1: {
+      component: Typography,
+      props: {
+        gutterBottom: true,
+        variant: 'h4',
+      },
+    },
+    h2: {
+      component: Typography,
+      props: { gutterBottom: true, variant: 'h6' },
+    },
+    h3: {
+      component: Typography,
+      props: { gutterBottom: true, variant: 'subtitle1' },
+    },
+    h4: {
+      component: Typography,
+      props: {
+        gutterBottom: true,
+        variant: 'caption',
+        paragraph: true,
+      },
+    },
+    p: {
+      component: Typography,
+      props: { paragraph: true },
+    },
+    a: { component: Link },
     li: {
-      component: withStyles(styles)(({ classes, ...props }) => (
-        <li className={classes.listItem}>
+      component: (props) => (
+        <Box component="li" sx={{ mt: 1 }}>
           <Typography component="span" {...props} />
-        </li>
-      )),
+        </Box>
+      ),
     },
   },
 };
 
-function Markdown(props) {
+export default function Markdown(props) {
   return <ReactMarkdown options={options} {...props} />;
 }
-
-export default Markdown;
