@@ -1,53 +1,62 @@
-import React from 'react';
-import classNames from 'classnames';
+import * as React from 'react';
 import PropTypes from 'prop-types';
-import MuiPaper from '@material-ui/core/Paper';
-import { capitalize } from '@material-ui/core/utils/helpers';
-import { withStyles } from '@material-ui/core/styles';
+import MuiPaper from '@mui/material/Paper';
+import { styled } from '@mui/material/styles';
 
-const styles = theme => ({
-  backgroundLight: {
-    backgroundColor: theme.palette.secondary.light,
-  },
-  backgroundMain: {
-    backgroundColor: theme.palette.secondary.main,
-  },
-  backgroundDark: {
-    backgroundColor: theme.palette.secondary.dark,
-  },
-  padding: {
-    padding: theme.spacing.unit,
-  },
-});
+const PaperRoot = styled(MuiPaper, {
+  shouldForwardProp: (prop) => prop !== 'background' && prop !== 'padding',
+})(({ theme }) => ({
+  variants: [
+    {
+      props: ({ padding }) => padding,
+      style: {
+        padding: theme.spacing(1),
+      },
+    },
+    {
+      props: { background: 'main' },
+      style: {
+        backgroundColor: theme.palette.secondary.main,
+      },
+    },
+    {
+      props: { background: 'light' },
+      style: {
+        backgroundColor: theme.palette.secondary.light,
+      },
+    },
+    {
+      props: { background: 'dark' },
+      style: {
+        backgroundColor: theme.palette.secondary.dark,
+      },
+    },
+  ],
+}));
 
 function Paper(props) {
-  const { background, classes, className, padding, ...other } = props;
+  const { background, classes, className, padding = false, ...other } = props;
+
   return (
-    <MuiPaper
-      elevation={0}
+    <PaperRoot
       square
-      className={classNames(
-        classes[`background${capitalize(background)}`],
-        {
-          [classes.padding]: padding,
-        },
-        className,
-      )}
+      elevation={0}
+      background={background}
+      padding={padding}
+      className={className}
       {...other}
     />
   );
 }
 
 Paper.propTypes = {
-  background: PropTypes.oneOf(['light', 'main', 'dark']),
-  classes: PropTypes.object.isRequired,
+  background: PropTypes.oneOf(['dark', 'light', 'main']).isRequired,
+  /**
+   * Override or extend the styles applied to the component.
+   */
+  classes: PropTypes.object,
   className: PropTypes.string,
   padding: PropTypes.bool,
 };
 
-Paper.defaultProps = {
-  background: 'light',
-  padding: false,
-};
-
-export default withStyles(styles)(Paper);
+export default Paper;
