@@ -1,41 +1,67 @@
-import React from 'react';
+import * as React from 'react';
 import PropTypes from 'prop-types';
-import { withStyles } from '@material-ui/core/styles';
-import { capitalize } from '@material-ui/core/utils/helpers';
-import MuiTypography from '@material-ui/core/Typography';
+import { styled } from '@mui/material/styles';
+import MuiTypography from '@mui/material/Typography';
 
-const styles = theme => ({
-  markedH2Center: {
+const markClassesMapping = {
+  center: {
+    h1: '',
+    h2: 'OnePirateTypography-markedH2Center',
+    h3: 'OnePirateTypography-markedH3Center',
+    h4: 'OnePirateTypography-markedH4Center',
+    h5: '',
+    h6: '',
+  },
+  left: {
+    h1: '',
+    h2: '',
+    h3: '',
+    h4: '',
+    h5: '',
+    h6: 'OnePirateTypography-markedH6Left',
+  },
+  none: {
+    h1: '',
+    h2: '',
+    h3: '',
+    h4: '',
+    h5: '',
+    h6: '',
+  },
+};
+
+const styles = ({ theme }) => ({
+  [`& .${markClassesMapping.center.h2}`]: {
     height: 4,
     width: 73,
     display: 'block',
-    margin: `${theme.spacing.unit}px auto 0`,
+    margin: `${theme.spacing(1)} auto 0`,
     backgroundColor: theme.palette.secondary.main,
   },
-  markedH3Center: {
+  [`& .${markClassesMapping.center.h3}`]: {
     height: 4,
     width: 55,
     display: 'block',
-    margin: `${theme.spacing.unit}px auto 0`,
+    margin: `${theme.spacing(1)} auto 0`,
     backgroundColor: theme.palette.secondary.main,
   },
-  markedH4Center: {
+  [`& .${markClassesMapping.center.h4}`]: {
     height: 4,
     width: 55,
     display: 'block',
-    margin: `${theme.spacing.unit}px auto 0`,
+    margin: `${theme.spacing(1)} auto 0`,
     backgroundColor: theme.palette.secondary.main,
   },
-  markedH6Left: {
+  [`& .${markClassesMapping.left.h6}`]: {
     height: 2,
     width: 28,
     display: 'block',
-    marginTop: theme.spacing.unit / 2,
+    marginTop: theme.spacing(0.5),
     background: 'currentColor',
   },
 });
 
-const headlineMapping = {
+const variantMapping = {
   h1: 'h1',
   h2: 'h1',
   h3: 'h1',
@@ -46,27 +72,47 @@ const headlineMapping = {
 };
 
 function Typography(props) {
-  const { children, classes, marked, variant, ...other } = props;
+  const { children, variant, marked = 'none', ...other } = props;
+
+  let markedClassName = '';
+  if (variant && variant in markClassesMapping[marked]) {
+    markedClassName = markClassesMapping[marked][variant];
+  }
 
   return (
-    <MuiTypography headlineMapping={headlineMapping} variant={variant} {...other}>
+    <MuiTypography variantMapping={variantMapping} variant={variant} {...other}>
       {children}
-      {marked ? (
-        <span className={classes[`marked${capitalize(variant) + capitalize(marked)}`]} />
-      ) : null}
+      {markedClassName ? <span className={markedClassName} /> : null}
     </MuiTypography>
   );
 }
 
 Typography.propTypes = {
+  /**
+   * The content of the component.
+   */
   children: PropTypes.node,
-  classes: PropTypes.object.isRequired,
-  marked: PropTypes.oneOf([false, 'center', 'left']),
-  variant: PropTypes.string,
+  marked: PropTypes.oneOf(['center', 'left', 'none']),
+  /**
+   * Applies the theme typography styles.
+   * @default 'body1'
+   */
+  variant: PropTypes.oneOf([
+    'body1',
+    'body2',
+    'button',
+    'caption',
+    'h1',
+    'h2',
+    'h3',
+    'h4',
+    'h5',
+    'h6',
+    'inherit',
+    'overline',
+    'subtitle1',
+    'subtitle2',
+  ]),
 };
 
-Typography.defaultProps = {
-  marked: false,
-};
-
-export default withStyles(styles)(Typography);
+export default styled(Typography)(styles);
