@@ -1,43 +1,46 @@
-import React from 'react';
+import * as React from 'react';
 import PropTypes from 'prop-types';
-import classNames from 'classnames';
-import { withStyles } from '@material-ui/core/styles';
+import { experimentalStyled as styled } from '@mui/material/styles';
+
 import Typography from '../components/Typography';
 
-const styles = theme => ({
-  root: {
-    padding: theme.spacing.unit * 2,
-  },
-  error: {
-    backgroundColor: theme.palette.error.xLight,
-    color: theme.palette.error.dark,
-  },
-  success: {
-    backgroundColor: theme.palette.success.xLight,
-    color: theme.palette.success.dark,
-  },
-});
+const Root = styled('div', {
+  shouldForwardProp: (prop) => prop !== 'error' && prop !== 'success',
+})(({ theme }) => ({
+  padding: theme.spacing(2),
+  variants: [
+    {
+      props: ({ error }) => error,
+      style: {
+        backgroundColor: theme.palette.error.light,
+        color: theme.palette.error.dark,
+      },
+    },
+    {
+      props: ({ success }) => success,
+      style: {
+        backgroundColor: theme.palette.success.light,
+        color: theme.palette.success.dark,
+      },
+    },
+  ],
+}));
 
 function FormFeedback(props) {
+  const { className, children, error, success, ...others } = props;
+
   return (
-    <div
-      className={classNames(
-        props.classes.root,
-        { [props.classes.error]: props.error, [props.classes.success]: props.success },
-        props.className,
-      )}
-    >
-      <Typography color="inherit">{props.children}</Typography>
-    </div>
+    <Root error={error} success={success} className={className} {...others}>
+      <Typography color="inherit">{children}</Typography>
+    </Root>
   );
 }
 
 FormFeedback.propTypes = {
   children: PropTypes.node,
-  classes: PropTypes.object.isRequired,
   className: PropTypes.string,
   error: PropTypes.bool,
   success: PropTypes.bool,
 };
 
-export default withStyles(styles)(FormFeedback);
+export default FormFeedback;
