@@ -1,60 +1,47 @@
-import React from 'react';
+import * as React from 'react';
 import PropTypes from 'prop-types';
-import classNames from 'classnames';
-import { withStyles } from '@material-ui/core/styles';
-import LayoutBody from '../components/LayoutBody';
+import { styled } from '@mui/material/styles';
 
-const styles = theme => ({
-  root: {
-    color: theme.palette.common.white,
-    position: 'relative',
-    display: 'flex',
-    alignItems: 'center',
-    [theme.breakpoints.up('sm')]: {
-      height: '80vh',
-      minHeight: 500,
-      maxHeight: 1300,
-    },
+import Container from '@mui/material/Container';
+import Box from '@mui/material/Box';
+
+const ProductHeroLayoutRoot = styled('section')(({ theme }) => ({
+  color: theme.palette.common.white,
+  position: 'relative',
+  display: 'flex',
+  alignItems: 'center',
+  [theme.breakpoints.up('sm')]: {
+    height: '80vh',
+    minHeight: 500,
+    maxHeight: 1300,
   },
-  layoutBody: {
-    marginTop: theme.spacing.unit * 3,
-    marginBottom: theme.spacing.unit * 14,
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'center',
-  },
-  backdrop: {
-    position: 'absolute',
-    left: 0,
-    right: 0,
-    top: 0,
-    bottom: 0,
-    backgroundColor: theme.palette.common.black,
-    opacity: 0.5,
-    zIndex: -1,
-  },
-  background: {
-    position: 'absolute',
-    left: 0,
-    right: 0,
-    top: 0,
-    bottom: 0,
-    backgroundSize: 'cover',
-    backgroundRepeat: 'no-repeat',
-    zIndex: -2,
-  },
-  arrowDown: {
-    position: 'absolute',
-    bottom: theme.spacing.unit * 4,
-  },
+}));
+
+const Background = styled('div')({
+  position: 'absolute',
+  left: 0,
+  right: 0,
+  top: 0,
+  bottom: 0,
+  backgroundSize: 'cover',
+  backgroundRepeat: 'no-repeat',
+  zIndex: -2,
 });
 
 function ProductHeroLayout(props) {
-  const { backgroundClassName, children, classes } = props;
+  const { sxBackground, children } = props;
 
   return (
-    <section className={classes.root}>
-      <LayoutBody className={classes.layoutBody} width="full">
+    <ProductHeroLayoutRoot>
+      <Container
+        sx={{
+          mt: 3,
+          mb: 14,
+          display: 'flex',
+          flexDirection: 'column',
+          alignItems: 'center',
+        }}
+      >
         <img
           src="/static/themes/onepirate/productHeroWonder.png"
           alt="wonder"
@@ -62,24 +49,39 @@ function ProductHeroLayout(props) {
           height="80"
         />
         {children}
-        <div className={classes.backdrop} />
-        <div className={classNames(classes.background, backgroundClassName)} />
-        <img
-          className={classes.arrowDown}
-          src="/static/themes/onepirate/productHeroArrowDown.png"
-          height="16"
-          width="12"
-          alt="arrow down"
+        <Box
+          sx={{
+            position: 'absolute',
+            left: 0,
+            right: 0,
+            top: 0,
+            bottom: 0,
+            backgroundColor: 'common.black',
+            opacity: 0.5,
+            zIndex: -1,
+          }}
         />
-      </LayoutBody>
-    </section>
+        <Background sx={sxBackground} />
+        <Box
+          component="img"
+          src="/static/themes/onepirate/productHeroArrowDown.png"
+          alt="arrow down"
+          sx={{ height: '16', width: '12', position: 'absolute', bottom: 32 }}
+        />
+      </Container>
+    </ProductHeroLayoutRoot>
   );
 }
 
 ProductHeroLayout.propTypes = {
-  backgroundClassName: PropTypes.string.isRequired,
-  children: PropTypes.node.isRequired,
-  classes: PropTypes.object.isRequired,
+  children: PropTypes.node,
+  sxBackground: PropTypes.oneOfType([
+    PropTypes.arrayOf(
+      PropTypes.oneOfType([PropTypes.func, PropTypes.object, PropTypes.bool]),
+    ),
+    PropTypes.func,
+    PropTypes.object,
+  ]),
 };
 
-export default withStyles(styles)(ProductHeroLayout);
+export default ProductHeroLayout;
