@@ -1,112 +1,105 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import { withStyles } from '@material-ui/core/styles';
-import Grid from '@material-ui/core/Grid';
-import Link from '@material-ui/core/Link';
-import LayoutBody from '../components/LayoutBody';
+import * as React from 'react';
+
+import Box from '@mui/material/Box';
+import Grid from '@mui/material/Grid';
+import Container from '@mui/material/Container';
 import Button from '../components/Button';
 import Typography from '../components/Typography';
 
-const styles = theme => ({
-  root: {
-    display: 'flex',
-    backgroundColor: theme.palette.secondary.light,
-    overflow: 'hidden',
-  },
-  layoutBody: {
-    marginTop: theme.spacing.unit * 10,
-    marginBottom: theme.spacing.unit * 15,
-    position: 'relative',
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'center',
-  },
-  item: {
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'center',
-    padding: `0px ${theme.spacing.unit * 5}px`,
-  },
-  title: {
-    marginBottom: theme.spacing.unit * 14,
-  },
-  number: {
-    fontSize: 24,
-    fontFamily: theme.typography.fontFamily,
-    color: theme.palette.secondary.main,
-    fontWeight: theme.typography.fontWeightMedium,
-  },
-  image: {
-    height: 55,
-    marginTop: theme.spacing.unit * 4,
-    marginBottom: theme.spacing.unit * 4,
-  },
-  curvyLines: {
-    pointerEvents: 'none',
-    position: 'absolute',
-    top: -180,
-    opacity: 0.7,
-  },
-  button: {
-    marginTop: theme.spacing.unit * 8,
-  },
-});
+const item = {
+  display: 'flex',
+  flexDirection: 'column',
+  alignItems: 'center',
+  px: 5,
+};
 
-function ProductHowItWorks(props) {
-  const { classes } = props;
+const number = {
+  fontSize: 24,
+  fontFamily: 'default',
+  color: 'secondary.main',
+  fontWeight: 'medium',
+};
 
+const image = {
+  height: 55,
+  my: 4,
+};
+
+function ProductHowItWorks() {
   return (
-    <section className={classes.root}>
-      <LayoutBody className={classes.layoutBody} width="large">
-        <img
+    <Box
+      component="section"
+      sx={{ display: 'flex', bgcolor: 'secondary.light', overflow: 'hidden' }}
+    >
+      <Container
+        sx={{
+          mt: 10,
+          mb: 15,
+          position: 'relative',
+          display: 'flex',
+          flexDirection: 'column',
+          alignItems: 'center',
+        }}
+      >
+        <Box
+          component="img"
           src="/static/themes/onepirate/productCurvyLines.png"
-          className={classes.curvyLines}
           alt="curvy lines"
+          sx={{
+            pointerEvents: 'none',
+            position: 'absolute',
+            top: -180,
+            opacity: 0.7,
+          }}
         />
-        <Typography variant="h4" marked="center" className={classes.title} component="h2">
+        <Typography variant="h4" marked="center" component="h2" sx={{ mb: 14 }}>
           How it works
         </Typography>
         <div>
-          <Grid container spacing={40}>
+          <Grid container spacing={5}>
             <Grid item xs={12} md={4}>
-              <div className={classes.item}>
-                <div className={classes.number}>1.</div>
-                <img
+              <Box sx={item}>
+                <Box sx={number}>1.</Box>
+                <Box
+                  component="img"
                   src="/static/themes/onepirate/productHowItWorks1.svg"
                   alt="suitcase"
-                  className={classes.image}
+                  sx={image}
                 />
                 <Typography variant="h5" align="center">
                   Appointment every Wednesday 9am.
                 </Typography>
-              </div>
+              </Box>
             </Grid>
             <Grid item xs={12} md={4}>
-              <div className={classes.item}>
-                <div className={classes.number}>2.</div>
-                <img
+              <Box sx={item}>
+                <Box sx={number}>2.</Box>
+                <Box
+                  component="img"
                   src="/static/themes/onepirate/productHowItWorks2.svg"
                   alt="graph"
-                  className={classes.image}
+                  sx={image}
                 />
                 <Typography variant="h5" align="center">
-                  First come, first served. Our offers are in limited quantities, so be quick.
+                  First come, first served. Our offers are in limited quantities, so
+                  be quick.
                 </Typography>
-              </div>
+              </Box>
             </Grid>
             <Grid item xs={12} md={4}>
-              <div className={classes.item}>
-                <div className={classes.number}>3.</div>
-                <img
+              <Box sx={item}>
+                <Box sx={number}>3.</Box>
+                <Box
+                  component="img"
                   src="/static/themes/onepirate/productHowItWorks3.svg"
                   alt="clock"
-                  className={classes.image}
+                  sx={image}
                 />
                 <Typography variant="h5" align="center">
                   {'New offers every week. New experiences, new surprises. '}
                   {'Your Sundays will no longer be alike.'}
                 </Typography>
-              </div>
+              </Box>
             </Grid>
           </Grid>
         </div>
@@ -114,20 +107,15 @@ function ProductHowItWorks(props) {
           color="secondary"
           size="large"
           variant="contained"
-          className={classes.button}
-          component={linkProps => (
-            <Link {...linkProps} href="/premium-themes/onepirate/sign-up" variant="button" />
-          )}
+          component="a"
+          href="/premium-themes/onepirate/sign-up/"
+          sx={{ mt: 8 }}
         >
           Get started
         </Button>
-      </LayoutBody>
-    </section>
+      </Container>
+    </Box>
   );
 }
 
-ProductHowItWorks.propTypes = {
-  classes: PropTypes.object.isRequired,
-};
-
-export default withStyles(styles)(ProductHowItWorks);
+export default ProductHowItWorks;
