@@ -1,110 +1,94 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import { withStyles } from '@material-ui/core/styles';
-import Grid from '@material-ui/core/Grid';
-import LayoutBody from '../components/LayoutBody';
+import * as React from 'react';
+
+import Box from '@mui/material/Box';
+import Grid from '@mui/material/Grid';
+import Container from '@mui/material/Container';
 import Typography from '../components/Typography';
 
-const styles = theme => ({
-  root: {
-    display: 'flex',
-    overflow: 'hidden',
-    backgroundColor: theme.palette.secondary.light,
-  },
-  layoutBody: {
-    marginTop: theme.spacing.unit * 15,
-    marginBottom: theme.spacing.unit * 30,
-    display: 'flex',
-    position: 'relative',
-  },
-  item: {
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'center',
-    padding: `0px ${theme.spacing.unit * 5}px`,
-  },
-  image: {
-    height: 55,
-  },
-  title: {
-    marginTop: theme.spacing.unit * 5,
-    marginBottom: theme.spacing.unit * 5,
-  },
-  curvyLines: {
-    pointerEvents: 'none',
-    position: 'absolute',
-    top: -180,
-  },
-});
+const item = {
+  display: 'flex',
+  flexDirection: 'column',
+  alignItems: 'center',
+  px: 5,
+};
 
-function ProductValues(props) {
-  const { classes } = props;
-
+function ProductValues() {
   return (
-    <section className={classes.root}>
-      <LayoutBody className={classes.layoutBody} width="large">
-        <img
+    <Box
+      component="section"
+      sx={{ display: 'flex', overflow: 'hidden', bgcolor: 'secondary.light' }}
+    >
+      <Container sx={{ mt: 15, mb: 30, display: 'flex', position: 'relative' }}>
+        <Box
+          component="img"
           src="/static/themes/onepirate/productCurvyLines.png"
-          className={classes.curvyLines}
           alt="curvy lines"
+          sx={{ pointerEvents: 'none', position: 'absolute', top: -180 }}
         />
-        <Grid container spacing={40}>
+        <Grid container spacing={5}>
           <Grid item xs={12} md={4}>
-            <div className={classes.item}>
-              <img
-                className={classes.image}
+            <Box sx={item}>
+              <Box
+                component="img"
                 src="/static/themes/onepirate/productValues1.svg"
                 alt="suitcase"
+                sx={{ height: 55 }}
               />
-              <Typography variant="h6" className={classes.title}>
+              <Typography variant="h6" sx={{ my: 5 }}>
                 The best luxury hotels
               </Typography>
               <Typography variant="h5">
-                {'From the latest trendy boutique hotel to the iconic palace with XXL pool'}
-                {', go for a mini-vacation just a few subway stops away from your home.'}
+                {
+                  'From the latest trendy boutique hotel to the iconic palace with XXL pool'
+                }
+
+                {
+                  ', go for a mini-vacation just a few subway stops away from your home.'
+                }
               </Typography>
-            </div>
+            </Box>
           </Grid>
           <Grid item xs={12} md={4}>
-            <div className={classes.item}>
-              <img
-                className={classes.image}
+            <Box sx={item}>
+              <Box
+                component="img"
                 src="/static/themes/onepirate/productValues2.svg"
                 alt="graph"
+                sx={{ height: 55 }}
               />
-              <Typography variant="h6" className={classes.title}>
+              <Typography variant="h6" sx={{ my: 5 }}>
                 New experiences
               </Typography>
               <Typography variant="h5">
-                {'Privatize a pool, take a Japanese bath or wake up in 900m2 of garden… '}
+                {
+                  'Privatize a pool, take a Japanese bath or wake up in 900m2 of garden… '
+                }
+
                 {'your Sundays will not be alike.'}
               </Typography>
-            </div>
+            </Box>
           </Grid>
           <Grid item xs={12} md={4}>
-            <div className={classes.item}>
-              <img
-                className={classes.image}
+            <Box sx={item}>
+              <Box
+                component="img"
                 src="/static/themes/onepirate/productValues3.svg"
                 alt="clock"
+                sx={{ height: 55 }}
               />
-              <Typography variant="h6" className={classes.title}>
+              <Typography variant="h6" sx={{ my: 5 }}>
                 Exclusive rates
               </Typography>
               <Typography variant="h5">
                 {'By registering, you will access specially negotiated rates '}
                 {'that you will not find anywhere else.'}
               </Typography>
-            </div>
+            </Box>
           </Grid>
         </Grid>
-      </LayoutBody>
-    </section>
+      </Container>
+    </Box>
   );
 }
 
-ProductValues.propTypes = {
-  classes: PropTypes.object.isRequired,
-};
-
-export default withStyles(styles)(ProductValues);
+export default ProductValues;
