@@ -1,173 +1,214 @@
-const bpmr = require('babel-plugin-module-resolver');
+// @ts-check
+const path = require('path');
 
-function resolvePath(sourcePath, currentFile, opts) {
-  if (sourcePath === 'markdown') {
-    const base = currentFile.substring(__dirname.length).slice(0, -3);
-    return `${__dirname}/docs/src/${base}/`;
-  }
+/**
+ * @typedef {import('@babel/core')} babel
+ */
 
-  return bpmr.resolvePath(sourcePath, currentFile, opts);
+const errorCodesPath = path.resolve(__dirname, './docs/public/static/error-codes.json');
+const missingError = process.env.MUI_EXTRACT_ERROR_CODES === 'true' ? 'write' : 'annotate';
+
+/**
+ * @param {string} relativeToBabelConf
+ * @returns {string}
+ */
+function resolveAliasPath(relativeToBabelConf) {
+  const resolvedPath = path.relative(process.cwd(), path.resolve(__dirname, relativeToBabelConf));
+  return `./${resolvedPath.replace('\\', '/')}`;
 }
 
-let defaultPresets;
+/** @type {babel.PluginItem[]} */
+const productionPlugins = [
+  ['babel-plugin-react-remove-properties', { properties: ['data-mui-test'] }],
+];
 
-// We release a ES version of Material-UI.
-// It's something that matches the latest official supported features of JavaScript.
-// Nothing more (stage-1, etc), nothing less (require, etc).
-if (process.env.BABEL_ENV === 'es') {
-  defaultPresets = [];
-} else {
-  defaultPresets = [
+/** @type {babel.ConfigFunction} */
+module.exports = function getBabelConfig(api) {
+  const useESModules = api.env(['regressions', 'modern', 'stable']);
+
+  const defaultAlias = {
+    '@mui/material': resolveAliasPath('./packages/mui-material/src'),
+    '@mui/docs': resolveAliasPath('./packages/mui-docs/src'),
+    '@mui/icons-material': resolveAliasPath(
+      `./packages/mui-icons-material/lib${useESModules ? '/esm' : ''}`,
+    ),
+    '@mui/lab': resolveAliasPath('./packages/mui-lab/src'),
+    '@mui/internal-markdown': resolveAliasPath('./packages/markdown'),
+    '@mui/styled-engine': resolveAliasPath('./packages/mui-styled-engine/src'),
+    '@mui/styled-engine-sc': resolveAliasPath('./packages/mui-styled-engine-sc/src'),
+    '@mui/styles': resolveAliasPath('./packages/mui-styles/src'),
+    '@mui/system': resolveAliasPath('./packages/mui-system/src'),
+    '@mui/private-theming': resolveAliasPath('./packages/mui-private-theming/src'),
+    '@mui/base': resolveAliasPath('./packages/mui-base/src'),
+    '@mui/utils': resolveAliasPath('./packages/mui-utils/src'),
+    '@mui/joy': resolveAliasPath('./packages/mui-joy/src'),
+    '@mui/internal-docs-utils': resolveAliasPath('./packages-internal/docs-utils/src'),
+    docs: resolveAliasPath('./docs'),
+    test: resolveAliasPath('./test'),
+  };
+
+  const presets = [
     [
       '@babel/preset-env',
       {
-        modules: ['modules', 'production-umd'].includes(process.env.BABEL_ENV) ? false : 'commonjs',
+        bugfixes: true,
+        browserslistEnv: process.env.BABEL_ENV || process.env.NODE_ENV,
+        debug: process.env.MUI_BUILD_VERBOSE === 'true',
+        modules: useESModules ? false : 'commonjs',
+        shippedProposals: api.env('modern'),
       },
     ],
+    [
+      '@babel/preset-react',
+      {
+        runtime: 'automatic',
+      },
+    ],
+    '@babel/preset-typescript',
   ];
-}
 
-const defaultAlias = {
-  '@material-ui/core': './packages/material-ui/src',
-  '@material-ui/icons': './packages/material-ui-icons/src',
-  '@material-ui/lab': './packages/material-ui-lab/src',
-  '@material-ui/styles': './packages/material-ui-styles/src',
-  '@material-ui/utils': './packages/material-ui-utils/src',
-  '@material-ui/system': './packages/material-ui-system/src',
-};
+  const usesAliases =
+    // in this config:
+    api.env(['coverage', 'development', 'test', 'benchmark']) ||
+    process.env.NODE_ENV === 'test' ||
+    // in webpack config:
+    api.env(['regressions']);
 
-module.exports = {
-  presets: defaultPresets.concat(['@babel/preset-react']),
-  plugins: [
-    ['@babel/plugin-proposal-class-properties', { loose: true }],
-    ['@babel/plugin-proposal-object-rest-spread', { loose: true }],
-    '@babel/plugin-transform-object-assign',
-    '@babel/plugin-transform-runtime',
-  ],
-  ignore: [/@babel[\\|/]runtime/],
-  env: {
-    test: {
-      sourceMaps: 'both',
-      plugins: [
-        [
-          'babel-plugin-module-resolver',
-          {
-            root: ['./'],
-            alias: defaultAlias,
-          },
+  const outFileExtension = '.js';
+
+  /** @type {babel.PluginItem[]} */
+  const plugins = [
+    [
+      'babel-plugin-macros',
+      {
+        muiError: {
+          errorCodesPath,
+          missingError,
+        },
+      },
+    ],
+    'babel-plugin-optimize-clsx',
+    [
+      '@babel/plugin-transform-runtime',
+      {
+        useESModules,
+        // any package needs to declare 7.4.4 as a runtime dependency. default is ^7.0.0
+        version: '^7.4.4',
+      },
+    ],
+    [
+      'babel-plugin-transform-react-remove-prop-types',
+      {
+        mode: 'unsafe-wrap',
+      },
+    ],
+    [
+      'transform-inline-environment-variables',
+      {
+        include: [
+          'MUI_VERSION',
+          'MUI_MAJOR_VERSION',
+          'MUI_MINOR_VERSION',
+          'MUI_PATCH_VERSION',
+          'MUI_PRERELEASE_LABEL',
+          'MUI_PRERELEASE_NUMBER',
         ],
-      ],
-    },
-    coverage: {
-      plugins: [
-        'babel-plugin-istanbul',
-        [
-          'babel-plugin-module-resolver',
-          {
-            root: ['./'],
-            alias: defaultAlias,
-          },
-        ],
-      ],
-    },
-    development: {
-      plugins: [
-        [
-          'babel-plugin-module-resolver',
-          {
-            alias: {
-              modules: './modules',
+      },
+    ],
+    ...(useESModules
+      ? [
+          [
+            '@mui/internal-babel-plugin-resolve-imports',
+            {
+              // Don't replace the extension when we're using aliases.
+              // Essentially only replace in production builds.
+              outExtension: usesAliases ? null : outFileExtension,
             },
-          },
-        ],
-      ],
+          ],
+        ]
+      : []),
+  ];
+
+  if (process.env.NODE_ENV === 'production') {
+    plugins.push(...productionPlugins);
+  }
+  if (process.env.NODE_ENV === 'test') {
+    plugins.push([
+      'babel-plugin-module-resolver',
+      {
+        alias: defaultAlias,
+        root: ['./'],
+      },
+    ]);
+  }
+
+  return {
+    assumptions: {
+      noDocumentAll: true,
     },
-    'docs-development': {
-      plugins: [
-        'babel-plugin-preval',
-        [
-          'babel-plugin-module-resolver',
-          {
-            alias: {
-              ...defaultAlias,
-              '@material-ui/docs': './packages/material-ui-docs/src',
-              docs: './docs',
-              modules: './modules',
-              pages: './pages',
+    presets,
+    plugins,
+    ignore: [/@babel[\\|/]runtime/], // Fix a Windows issue.
+    overrides: [
+      {
+        exclude: /\.test\.(js|ts|tsx)$/,
+        plugins: ['@babel/plugin-transform-react-constant-elements'],
+      },
+      {
+        test: /(\.test\.[^.]+$|\.test\/)/,
+        plugins: [['@mui/internal-babel-plugin-resolve-imports', false]],
+      },
+    ],
+    env: {
+      coverage: {
+        plugins: [
+          'babel-plugin-istanbul',
+          [
+            'babel-plugin-module-resolver',
+            {
+              root: ['./'],
+              alias: defaultAlias,
             },
-            transformFunctions: ['require', 'require.context'],
-            resolvePath,
-          },
+          ],
         ],
-      ],
-    },
-    'docs-production': {
-      plugins: [
-        'babel-plugin-preval',
-        [
-          'babel-plugin-module-resolver',
-          {
-            alias: {
-              ...defaultAlias,
-              '@material-ui/docs': './packages/material-ui-docs/src',
-              docs: './docs',
-              modules: './modules',
-              pages: './pages',
+      },
+      development: {
+        plugins: [
+          [
+            'babel-plugin-module-resolver',
+            {
+              alias: {
+                ...defaultAlias,
+                modules: './modules',
+              },
+              root: ['./'],
             },
-            transformFunctions: ['require', 'require.context'],
-            resolvePath,
-          },
+          ],
         ],
-        'transform-react-constant-elements',
-        'transform-dev-warning',
-        ['react-remove-properties', { properties: ['data-mui-test'] }],
-        ['transform-react-remove-prop-types', { mode: 'remove' }],
-      ],
-    },
-    es: {
-      plugins: [
-        'transform-react-constant-elements',
-        'transform-dev-warning',
-        ['react-remove-properties', { properties: ['data-mui-test'] }],
-        [
-          'transform-react-remove-prop-types',
-          {
-            mode: 'unsafe-wrap',
-          },
+      },
+      test: {
+        sourceMaps: 'both',
+        plugins: [
+          [
+            'babel-plugin-module-resolver',
+            {
+              root: ['./'],
+              alias: defaultAlias,
+            },
+          ],
         ],
-      ],
-      // It's most likely a babel bug.
-      // We are using this ignore option in the CLI command but that has no effect.
-      ignore: ['**/*.test.js'],
-    },
-    production: {
-      plugins: [
-        'transform-react-constant-elements',
-        'transform-dev-warning',
-        ['react-remove-properties', { properties: ['data-mui-test'] }],
-        [
-          'transform-react-remove-prop-types',
-          {
-            mode: 'unsafe-wrap',
-          },
+      },
+      benchmark: {
+        plugins: [
+          ...productionPlugins,
+          [
+            'babel-plugin-module-resolver',
+            {
+              alias: defaultAlias,
+            },
+          ],
         ],
-      ],
-      // It's most likely a babel bug.
-      // We are using this ignore option in the CLI command but that has no effect.
-      ignore: ['**/*.test.js'],
+      },
     },
-    'production-umd': {
-      plugins: [
-        'transform-react-constant-elements',
-        'transform-dev-warning',
-        ['react-remove-properties', { properties: ['data-mui-test'] }],
-        [
-          'transform-react-remove-prop-types',
-          {
-            mode: 'unsafe-wrap',
-          },
-        ],
-      ],
-    },
-  },
+  };
 };
