@@ -1,20 +1,19 @@
 /* eslint-disable no-console */
+const path = require('path');
+const gm = require('gm');
 
-import path from 'path';
-import gm from 'gm';
-
-const SIZES = [48, 70, 96, 150, 152, 192, 256, 310, 384, 512];
-const INPUT_ICON = path.join(__dirname, '../../static/logo.png');
-const OUTPUT_DIR = path.join(__dirname, '../../static/icons');
+const SIZES = [48, 96, 180, 192, 256, 384, 512];
+const INPUT_ICON = path.join(__dirname, '../public/static/logo.png');
+const OUTPUT_DIR = path.join(__dirname, '../public/static/icons');
 
 console.log('Generating Icons');
 
 const promises = SIZES.map(
-  size =>
+  (size) =>
     new Promise((resolve, reject) => {
       gm(INPUT_ICON)
         .resize(size, size)
-        .write(path.join(OUTPUT_DIR, `${size}x${size}.png`), err => {
+        .write(path.join(OUTPUT_DIR, `${size}x${size}.png`), (err) => {
           if (err) {
             reject(err);
             return;
@@ -26,9 +25,9 @@ const promises = SIZES.map(
     }),
 );
 
-Promise.all(promises).catch(err => {
+Promise.all(promises).catch((err) => {
   setTimeout(() => {
     console.log(err);
     throw err;
-  }, 0);
+  });
 });
