@@ -1,9 +1,24 @@
-export const CODE_VARIANTS = {
+const CODE_VARIANTS = {
   JS: 'JS',
-  HOOK: 'HOOK',
+  TS: 'TS',
 };
 
-export const ACTION_TYPES = {
-  THEME_CHANGE: 'THEME_CHANGE',
-  OPTIONS_CHANGE: 'OPTIONS_CHANGE',
+const CODE_STYLING = {
+  SYSTEM: 'MUI System',
+  TAILWIND: 'Tailwind',
+  CSS: 'CSS',
+};
+
+// Valid languages to use in production
+const LANGUAGES_LABEL = [
+  {
+    code: 'en',
+    text: 'English',
+  },
+];
+
+module.exports = {
+  CODE_VARIANTS,
+  LANGUAGES_LABEL,
+  CODE_STYLING,
 };
