@@ -1,14 +1,25 @@
-import React from 'react';
+import * as React from 'react';
 import PropTypes from 'prop-types';
-import defer from './defer';
+
 import Button from '../components/Button';
+import defer from './defer';
 
 function FormButton(props) {
   const { disabled, mounted, ...others } = props;
-  return <Button disabled={!mounted || disabled} type="submit" variant="contained" {...others} />;
+  return (
+    <Button
+      disabled={!mounted || !!disabled}
+      type="submit"
+      variant="contained"
+      {...others}
+    />
+  );
 }
 
 FormButton.propTypes = {
+  /**
+   * If `true`, the component is disabled.
+   */
   disabled: PropTypes.bool,
   mounted: PropTypes.bool,
 };
